requirements.txt
aiohappyeyeballs==2.6.1
aiosignal==1.4.0
attrs==25.4.0
backoff==1.11.1
certifi==2026.2.25
charset-normalizer==3.4.5
demucs>=4.0.0
frozenlist==1.8.0
idna==3.11
lyricsgenius>=3.0.0
openai-whisper>=20231117
propcache==0.4.1
python-dotenv==1.2.2
requests==2.32.5
spotipy==2.26.0
typing_extensions==4.15.0
urllib3==2.6.3
yarl==1.23.0
yt-dlp>=2024.1.0
demucs>=4.0.0
openai-whisper>=20231117
lyricsgenius>=3.0.0
